Small bugfix

+++ src/utils/syncDirs.ts
@@ -21,30 +21,33 @@ export const syncDirs = (
       recursive: true,
     },
     (event, fileChangedPath) => {
-      onStartSync?.();
-      const fileSrcDir = path.dirname(fileChangedPath);
       const fileName = path.basename(fileChangedPath);
-      const fileOutDir = fileSrcDir.replace(path.normalize(srcDir), outDir);
-      console.log(`  ${coloredString(`File ${fileChangedPath} ${event}.`)}`);
-      if (event === "remove") {
-        const pathToRemove = getPath(`${fileOutDir}/${fileName}`);
-        const fileTransformer = transformers?.filter((x) =>
-          x.fileRegex.test(fileName),
-        );
-        if (fileTransformer.length > 0) {
-          fileTransformer.forEach(async (x) => {
-            if (x.fileRegex.test(fileChangedPath)) {
-              const finalPathToRemove =
-                x.pathTransformer?.(pathToRemove) ?? pathToRemove;
-              fs.rmSync(finalPathToRemove, { force: true, recursive: true });
-            }
-          });
-        } else fs.rmSync(pathToRemove, { force: true, recursive: true });
-      } else {
-        copyFile(fileSrcDir, fileName, fileOutDir, transformers, omit);
-      }
+      // Its a directory
+      if (fileChangedPath !== fileName) {
+        onStartSync?.();
+        const fileSrcDir = path.dirname(fileChangedPath);
+        const fileOutDir = fileSrcDir.replace(path.normalize(srcDir), outDir);
+        console.log(`  ${coloredString(`File ${fileChangedPath} ${event}.`)}`);
+        if (event === "remove") {
+          const pathToRemove = getPath(`${fileOutDir}/${fileName}`);
+          const fileTransformer = transformers?.filter((x) =>
+            x.fileRegex.test(fileName),
+          );
+          if (fileTransformer.length > 0) {
+            fileTransformer.forEach(async (x) => {
+              if (x.fileRegex.test(fileChangedPath)) {
+                const finalPathToRemove =
+                  x.pathTransformer?.(pathToRemove) ?? pathToRemove;
+                fs.rmSync(finalPathToRemove, { force: true, recursive: true });
+              }
+            });
+          } else fs.rmSync(pathToRemove, { force: true, recursive: true });
+        } else {
+          copyFile(fileSrcDir, fileName, fileOutDir, transformers, omit);
+        }
 
-      onEndSync?.(event, fileChangedPath);
+        onEndSync?.(event, fileChangedPath);
+      }
     },
   );
 };
